Move the conditional menu code into a separate file.
